+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Font class included in public API.
+-   Glyphmap with dynamic font demo.
+-   Glyphset internal class for use in managing textures for Glyphmap - improve Glyphmap refresh performance.
+
+### Changed
+
+-   Glyphmap uses Glyphset internally and in renderers.
+
+### Removed
+
+-   Remove extraneous glyph data from glyphmap instance; use texture keys to reconstitute glyph data when refreshing.
+
 ## [0.2.0] - 2022-01-08
 
 ### Added